Find bugs related to rollback transactions and solve them. Moreover, make some test codes correct.

+++ onthego.database/src/main/java/onthego/database/core/tablespace/manager/SingleTablespaceManager.java
@@ -102,7 +102,7 @@ public class SingleTablespaceManager implements TablespaceManager {
 			io.writeLong(tsHeader.getFirstFreeBlockPos());
 			io.writeLong(tsHeader.getTableRootPos());
 			io.writeLong(tsHeader.getTableMetaInfoPos());
-			io.writeLong(tsHeader.getRecordCount());
+			io.writeInt(tsHeader.getRecordCount());
 		} catch(Exception ioe) {
 			throw new TablespaceManagerException(ioe);
 		}
